move `blogData` helper to react_props
